Add label access flag to a field value
Usually, it is enough to use field code to access record values.
However, it may help to be able to access via label value if an app
is designed to use unique labels. In this case, we can use labels
as a key to access values although it may pay a little performance
cost.

Set a flag like the following 3rd param.
  aexlib.kintone.App.getApp(id, fields, {labelAccess:true});
Or, it can set later like:
  var app = aexlib.kintone.App.getApp();
  app.labelAccess(true);

To access via label, fields are required. So, it may be required
to fetch fields first.
  app.fetchFields().then(function() {
    // access to fields...
  });